requirements.txt
accelerate
numpy
Pillow
requests
tqdm
diffusers
transformers
